server/requirements.txt
Flask
Flask-SQLAlchemy
Flask-JWT-Extended
Flask-CORS
Flask-Bcrypt
python-dotenv 
pypdf
pdfminer.six